Move the CopyItem class in tools.h
This way, other binaries will be able to use it.
