+++ src/tools.h
@@ -27,6 +27,10 @@
 #include <stdexcept>
 #include <sstream>
 
+#include <zim/writer/contentProvider.h>
+#include <zim/writer/item.h>
+#include <zim/item.h>
+
 /* Formatter for std::exception what() message:
  * throw std::runtime_error(
  *   Formatter() << "zimwriterfs: Unable to read" << filename << ": " << strerror(errno));
@@ -60,6 +64,62 @@ typedef struct html_link
     std::string attribute;
     std::string link;
 } html_link;
+
+// Few helper class to help copy a item from a archive to another one.
+class ItemProvider : public zim::writer::ContentProvider
+{
+    zim::Item item;
+    bool feeded;
+  public:
+    ItemProvider(zim::Item item)
+      : item(item),
+        feeded(false)
+    {}
+
+    zim::size_type getSize() const {
+      return item.getSize();
+    }
+
+    zim::Blob feed() {
+      if (feeded) {
+        return zim::Blob();
+      }
+      feeded = true;
+      return item.getData();
+    }
+};
+
+
+class CopyItem : public zim::writer::Item         //Article class that will be passed to the zimwriter. Contains a zim::Article class, so it is easier to add a
+{
+    //article from an existing ZIM file.
+    zim::Item item;
+
+  public:
+    explicit CopyItem(const zim::Item item):
+      item(item)
+    {}
+
+    virtual std::string getPath() const
+    {
+        return item.getPath();
+    }
+
+    virtual std::string getTitle() const
+    {
+        return item.getTitle();
+    }
+
+    virtual std::string getMimeType() const
+    {
+        return item.getMimetype();
+    }
+
+    std::unique_ptr<zim::writer::ContentProvider> getContentProvider() const
+    {
+       return std::unique_ptr<zim::writer::ContentProvider>(new ItemProvider(item));
+    }
+};
 
 std::string getMimeTypeForFile(const std::string& basedir, const std::string& filename);
 std::string getNamespaceForMimeType(const std::string& mimeType, bool uniqueNamespace);
@@ -101,5 +161,4 @@ int adler32(const std::string& buf);
 //Converts the %20 to space.Essential for comparing URLs.
 std::string normalize_link(const std::string& input, const std::string& baseUrl);
 
-
-#endif  // OPENZIM_TOOLS_H
+#endif  // OPENZIM_TOOLS_H
