+++ src/zimdiff.cpp
@@ -21,14 +21,14 @@
 #include <sstream>
 #include <vector>
 #include <zim/writer/creator.h>
-#include <zim/writer/contentProvider.h>
-#include <zim/writer/item.h>
 #include <zim/blob.h>
 #include <zim/archive.h>
 #include <zim/item.h>
 #include <list>
 #include <algorithm>
 #include <sstream>
+
+#include "tools.h"
 
 #include "version.h"
 
@@ -39,60 +39,6 @@ std::string NumberToString(int number)
   return ss.str();
 }
 
-class ItemProvider : public zim::writer::ContentProvider
-{
-    zim::Item item;
-    bool feeded;
-  public:
-    ItemProvider(zim::Item item)
-      : item(item),
-        feeded(false)
-    {}
-
-    zim::size_type getSize() const {
-      return item.getSize();
-    }
-
-    zim::Blob feed() {
-      if (feeded) {
-        return zim::Blob();
-      }
-      feeded = true;
-      return item.getData();
-    }
-};
-
-
-class CopyItem : public zim::writer::Item         //Article class that will be passed to the zimwriter. Contains a zim::Article class, so it is easier to add a
-{
-    //article from an existing ZIM file.
-    zim::Item item;
-
-  public:
-    explicit CopyItem(const zim::Item item):
-      item(item)
-    {}
-
-    virtual std::string getPath() const
-    {
-        return item.getPath();
-    }
-
-    virtual std::string getTitle() const
-    {
-        return item.getTitle();
-    }
-
-    virtual std::string getMimeType() const
-    {
-        return item.getMimetype();
-    }
-
-    std::unique_ptr<zim::writer::ContentProvider> getContentProvider() const
-    {
-       return std::unique_ptr<zim::writer::ContentProvider>(new ItemProvider(item));
-    }
-};
 
 void create(const std::string& filename_1, const std::string& filename_2, const std::string& outpath)
 {
